geolocation update to 2x

+++ web/modules/contrib/geolocation/src/Plugin/Field/FieldFormatter/GeolocationTokenFormatter.php
@@ -29,7 +29,10 @@ class GeolocationTokenFormatter extends FormatterBase {
    */
   public static function defaultSettings() {
     $settings = [];
-    $settings['tokenized_text'] = '';
+    $settings['tokenized_text'] = [
+      'value' => '',
+      'format' => filter_default_format(),
+    ];
     $settings += parent::defaultSettings();
 
     return $settings;
@@ -42,11 +45,17 @@ class GeolocationTokenFormatter extends FormatterBase {
     $settings = $this->getSettings();
 
     $element['tokenized_text'] = [
-      '#type' => 'textarea',
+      '#type' => 'text_format',
       '#title' => $this->t('Tokenized text'),
       '#description' => $this->t('Enter any text or HTML to be shown for each value. Tokens will be replaced as available. The "token" module greatly expands the number of available tokens as well as provides a comfortable token browser.'),
-      '#default_value' => $settings['tokenized_text'],
     ];
+    if (!empty($settings['tokenized_text']['value'])) {
+      $form['tokenized_text']['#default_value'] = $settings['tokenized_text']['value'];
+    }
+
+    if (!empty($settings['info_text']['format'])) {
+      $form['tokenized_text']['#format'] = $settings['tokenized_text']['format'];
+    }
 
     $element['token_help'] = $this->getTokenHelp();
 
@@ -60,14 +69,19 @@ class GeolocationTokenFormatter extends FormatterBase {
     $settings = $this->getSettings();
 
     $summary = [];
-    $summary[] = $this->t('Tokenized Text: %text', [
-      '%text' => Unicode::truncate(
-        $settings['tokenized_text'],
-        100,
-        TRUE,
-        TRUE
-      ),
-    ]);
+    if (
+      !empty($settings['tokenized_text']['value'])
+      && !empty($settings['tokenized_text']['format'])
+    ) {
+      $summary[] = $this->t('Tokenized Text: %text', [
+        '%text' => Unicode::truncate(
+          check_markup($settings['tokenized_text']['value'], $settings['tokenized_text']['format']),
+          100,
+          TRUE,
+          TRUE
+        ),
+      ]);
+    }
 
     return $summary;
   }
@@ -83,13 +97,26 @@ class GeolocationTokenFormatter extends FormatterBase {
     $elements = [];
     foreach ($items as $delta => $item) {
       $token_context['geolocation_current_item'] = $item;
-      $tokenized_text = \Drupal::token()->replace($this->getSetting('tokenized_text'), $token_context, [
-        'callback' => [$this, 'geolocationItemTokens'],
-        'clear' => TRUE,
-      ]);
-      $elements[$delta] = [
-        '#markup' => $tokenized_text,
-      ];
+
+      $tokenized_text = $this->getSetting('tokenized_text');
+
+      if (
+        !empty($tokenized_text['value'])
+        && !empty($tokenized_text['format'])
+      ) {
+        $elements[$delta] = [
+          '#type' => 'processed_text',
+          '#text' => \Drupal::token()->replace(
+            $tokenized_text['value'],
+            $token_context,
+            [
+              'callback' => [$this, 'geolocationItemTokens'],
+              'clear' => TRUE,
+            ]
+          ),
+          '#format' => $tokenized_text['format'],
+        ];
+      }
     }
 
     return $elements;
